+++ src/compas_viewer/components/__init__.py
@@ -1,7 +1,8 @@
 from .button import Button
 from .combobox import ComboBox
 from .combobox import ViewModeAction
-from .dialog import CameraSettingsDialog
+from .camerasetting import CameraSettingsDialog
+from .objectsetting import ObjectSettingDialog
 from .slider import Slider
 from .treeform import Treeform
 from .sceneform import Sceneform
@@ -10,6 +11,7 @@ __all__ = [
     "Button",
     "ComboBox",
     "CameraSettingsDialog",
+    "ObjectSettingDialog",
     "Renderer",
     "Slider",
     "Treeform",
