+++ src/compas_viewer/components/camerasetting.py
@@ -71,70 +71,3 @@ class CameraSettingsDialog(QDialog, Base):
             self.spin_boxes["Camera_Position_Z"].spinbox.value(),
         )
         self.accept()
-
-
-class ObjectInfoDialog(QDialog, Base):
-    """
-    A dialog for displaying and updating object settings in Qt applications.
-    This dialog allows users to modify object properties such as line width, point size, and opacity,
-    and applies these changes dynamically.
-
-    Attributes
-    ----------
-    layout : QVBoxLayout
-        The layout of the dialog.
-    spin_boxes : dict
-        Dictionary containing spin boxes for adjusting object properties.
-    update_button : QPushButton
-        Button to apply changes to the selected objects.
-
-    Methods
-    -------
-    update()
-        Updates the properties of selected objects and closes the dialog.
-
-    Example
-    -------
-    >>> dialog = ObjectInfoDialog()
-    >>> dialog.exec()
-    """
-
-    def __init__(self) -> None:
-        super().__init__()
-
-        self.setWindowTitle("Object Settings")
-        self.layout = QVBoxLayout(self)
-
-        coordinates = {}
-        for obj in self.viewer.scene.objects:
-            if obj.is_selected:
-                new_coordinates = {
-                    "Name": [("label", str(obj.name))],
-                    "Parent": [("label", str(obj.parent))],
-                    # TODO: check _color attr ("double_edit", "G", obj.linecolor[0].g, 0, 1),
-                    "Point_Color": [("color_combobox", obj, "pointcolor")],
-                    "Line_Color": [("color_combobox", obj, "linecolor")],
-                    "Face_Color": [("color_combobox", obj, "facecolor")],
-                    "Line_Width": [("double_edit", "", obj.linewidth, 0.0, 10.0)],
-                    "Point_Size": [("double_edit", "", obj.pointsize, 0.0, 10.0)],
-                    "Opacity": [("double_edit", "", obj.opacity, 0.0, 1.0)],
-                }
-                coordinates.update(new_coordinates)
-
-        object_info_layout, self.spin_boxes = base_layout(coordinates)
-
-        self.layout.addLayout(object_info_layout)
-
-        self.update_button = QPushButton("Update Object", self)
-        self.update_button.clicked.connect(self.update)
-        self.layout.addWidget(self.update_button)
-
-    def update(self) -> None:
-        for obj in self.viewer.scene.objects:
-            if obj.is_selected:
-                obj.linewidth = self.spin_boxes["Line_Width_"].spinbox.value()
-                obj.pointsize = self.spin_boxes["Point_Size_"].spinbox.value()
-                obj.opacity = self.spin_boxes["Opacity_"].spinbox.value()
-                obj.update()
-
-        self.accept()
