+++ src/compas_viewer/config.py
@@ -17,7 +17,6 @@ from compas_viewer.commands import clear_scene_cmd
 from compas_viewer.commands import deselect_all_cmd
 from compas_viewer.commands import load_scene_cmd
 from compas_viewer.commands import obj_settings_cmd
-from compas_viewer.commands import object_info_qmsg_cmd
 from compas_viewer.commands import pan_view_cmd
 from compas_viewer.commands import rotate_view_cmd
 from compas_viewer.commands import save_scene_cmd
@@ -383,7 +382,6 @@ class Config(ConfigBase):
             toggle_toolbar_cmd,
             zoom_selected_cmd,
             zoom_view_cmd,
-            object_info_qmsg_cmd,
             obj_settings_cmd,
         ]
     )
